updated functions responsible for link processing and checking.
Now the link is checked all the time while typing, and when the link is correct, the button turns blue.

The verify_url function takes different parameters; now the url is taken inside the function and not accepted as an argument. A silent argument has been added which allows you to block the display of messages if the link is invalid.

In the process_url function, instead of fetch, I use XMLHttpRequest because I am used to XHR and I work better with it.

A function has been added that allows you to reset the input and displayed messages on the page.

A function has been added that copies the shortened link when the user clicks on it.

And of course I added some minor styling updates.

That's it for frontend updates.

In main.py I added a mechanism that checks whether the connection to the database has been established correctly, if not, the application does not start.